# app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bilgi Evim Eğitim Yönetim Sistemi API",
    "description": "Bilgi Evim Eğitim yönetim sistemi için oluşturulmuş API'ler.",
    "version": "1.0.0"
  },
  "paths": {
    "/admin/dashboard": {
      "get": {
        "tags": ["Admin Routes"],
        "summary": "Admin Dashboard",
        "description": "Admin'in ana sayfasına erişim sağlar.",
        "security": [{"jwtAuth": []}],
        "responses": {
          "200": {
            "description": "Başarılı",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Admin dashboard"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Yetkisiz erişim"
          }
        }
      }
    },
    "/admin/{admin_id}": {
      "put": {
        "tags": ["Admin Routes"],
        "summary": "Admin Bilgi Güncelleme",
        "description": "Admin bilgilerini günceller.",
        "parameters": [
          {
            "name": "admin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [{"jwtAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {"type": "string"},
                  "password": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Başarılı",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Admin profile updated successfully"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Yetkisiz erişim"
          }
        }
      }
    },
    "/admin/student-list": {
      "get": {
        "tags": ["Admin Routes"],
        "summary": "Tüm Öğrencileri Getir",
        "description": "Sistemde kayıtlı tüm öğrencileri getirir.",
        "security": [{"jwtAuth": []}],
        "responses": {
          "200": {
            "description": "Başarılı",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {"type": "integer"},
                      "name": {"type": "string"},
                      "lastname": {"type": "string"},
                      "school_number": {"type": "integer"},
                      "tc": {"type": "string"},
                      "class_id": {"type": "integer"}
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Yetkisiz erişim"
          }
        }
      }
    },
    "/admin/student/{student_id}": {
      "put": {
        "tags": ["Admin Routes"],
        "summary": "Öğrenci Profili Güncelleme",
        "description": "Öğrenci profil bilgilerini günceller.",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [{"jwtAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {"type": "string"},
                  "lastname": {"type": "string"},
                  "tc": {"type": "string"},
                  "school_number": {"type": "integer"},
                  "grade": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Başarılı",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Student updated successfully"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Yetkisiz erişim"
          }
        }
      },
      "delete": {
        "tags": ["Admin Routes"],
        "summary": "Öğrenci Sil",
        "description": "Belirtilen öğrenci ID'sine göre öğrenciyi siler.",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [{"jwtAuth": []}],
        "responses": {
          "200": {
            "description": "Başarılı",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Yetkisiz erişim"
          }
        }
      }
    },
    "/admin/create-student": {
      "post": {
        "tags": ["Admin Routes"],
        "summary": "Öğrenci Kayıt İşlemi",
        "description": "Yeni bir öğrenci kaydı oluşturur.",
        "security": [{"jwtAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {"type": "string"},
                  "lastname": {"type": "string"},
                  "tc": {"type": "string"},
                  "school_number": {"type": "integer"},
                  "password": {"type": "string"},
                  "class_id": {"type": "integer"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Başarılı",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Student registered successfully"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Yetkisiz erişim"
          }
        }
      }
    },
    "/auth/login/admin": {
      "post": {
        "tags": ["Auth"],
        "summary": "Admin Login",
        "description": "Admin'in giriş işlemini sağlar.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "admin"
                  },
                  "password": {
                    "type": "string",
                    "example": "admin"
                  }
                },
                "required": ["name", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Başarılı",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1..."
                    },
                    "role": {
                      "type": "string",
                      "example": "admin"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Geçersiz kimlik bilgileri",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid credentials"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "jwtAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Admin": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "name": {"type": "string"},
          "role": {"type": "string"}
        }
      },
      "Student": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "name": {"type": "string"},
          "lastname": {"type": "string"},
          "school_number": {"type": "integer"},
          "tc": {"type": "string"},
          "grade": {"type": "string"}
        }
      },
      "Teacher": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "name": {"type": "string"},
          "lastname": {"type": "string"},
          "teacher_number": {"type": "integer"},
          "subject": {"type": "string"}
        }
      }
    }
  }
}